Introspect unknown Bearer tokens
The openidc cache does not know of any tokens which have not
been issued via its module, even if the token would be valid to
the authorization server. Therefore, it would start a new oauth
delegation process, i.e. API requests would get a 302 redirection
response and interrupt due to CORS issues.

To solve this issue openidc has to be configured to act as a
resource server as documented here:

  https://github.com/zmartzone/mod_auth_openidc/wiki/OAuth-2.0-Resource-Server

With this, the module will ask the authorization server to introspect an
incoming Bearer token which is unknown. This is useful for clients
 already having  a valid access token valid for accessing the STA.

httpd configuration distinguishes between both OAuth rqeuests (having
a Bearer token) and all other requests. When a Bearer token is present,
httpd is configured to introspect the token via authorization server, or
otherwise to start the authorization process where the user has to take
active part in.